refactor(rdp): inline service runtime, drop bestool-winsvc crate
alertd already uses define_windows_service! inside its own crate, so macro
hygiene keeps the unsafe extern tokens in the windows-service crate's span
— the workspace's unsafe_code = forbid doesn't reach them. Follow the same
pattern here and merge the ServiceMain dispatcher into rdp/service.rs, so
we don't ship an extra workspace crate just for the FFI shim.
