+++ crates/bestool/Cargo.toml
@@ -83,7 +83,6 @@ zip = { version = "8.2.0", optional = true, default-features = false, features =
 zmq = { version = "0.10.0", optional = true }
 
 [target.'cfg(windows)'.dependencies]
-bestool-winsvc = { version = "0.1.0", path = "../winsvc", optional = true }
 tauri-winrt-notification = { version = "0.7.2", optional = true }
 windows-acl = { version = "0.3.0", optional = true }
 windows-env = { version = "0.2.0", optional = true }
@@ -103,7 +102,7 @@ caddy = ["download", "dep:tera"]
 completions = ["dep:clap_complete", "dep:clap_complete_nushell"]
 crypto = ["dep:algae-cli", "dep:blake3", "dep:merkle_hash"]
 file = ["dep:blake3", "dep:indicatif", "dep:tokio-util"]
-rdp = ["dep:bestool-winsvc", "dep:duct", "dep:privilege", "dep:quick-xml", "dep:tauri-winrt-notification", "dep:windows-service"]
+rdp = ["dep:duct", "dep:privilege", "dep:quick-xml", "dep:tauri-winrt-notification", "dep:windows-service"]
 self-update = ["download", "dep:upgrade", "dep:windows-env"]
 ssh = ["dep:dirs", "dep:duct", "dep:fs4", "dep:ssh-key", "dep:privilege", "dep:windows-acl"]
