--- crates/winsvc/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "bestool-winsvc"
-version = "0.1.0"
-edition = "2024"
-
-authors = ["BES Developers <contact@bes.au>"]
-license = "GPL-3.0-or-later"
-description = "Windows Service runtime glue for bestool"
-repository = "https://github.com/beyondessential/bestool"
-publish = false
-
-[target.'cfg(windows)'.dependencies]
-tokio = { workspace = true, features = ["sync"] }
-windows-service = "0.8.0"
-
-[lints.rust]
-rust_2018_idioms = { level = "deny", priority = -1 }
