--- crates/winsvc/src/lib.rs
@@ -1,124 +1,0 @@
-//! Minimal Windows Service runtime for bestool.
-//!
-//! Lives in its own crate because the Service Control Manager handshake
-//! (`StartServiceCtrlDispatcher` / `ServiceMain`) requires one `unsafe extern
-//! "system" fn`, which is generated by the `define_windows_service!` macro
-//! from the `windows-service` crate. The bestool workspace forbids `unsafe`,
-//! so the glue is isolated here.
-//!
-//! `run(name, entry)` blocks until the SCM signals Stop or Shutdown; until
-//! then `entry` is invoked with a shutdown receiver that fires when Windows
-//! asks the service to exit.
-
-#![cfg_attr(not(windows), allow(dead_code))]
-
-#[cfg(windows)]
-mod imp {
-	use std::{ffi::OsString, sync::OnceLock, time::Duration};
-
-	use tokio::sync::watch;
-	use windows_service::{
-		Result, define_windows_service,
-		service::{
-			ServiceControl, ServiceControlAccept, ServiceExitCode, ServiceState, ServiceStatus,
-			ServiceType,
-		},
-		service_control_handler::{self, ServiceControlHandlerResult},
-		service_dispatcher,
-	};
-
-	const SERVICE_TYPE: ServiceType = ServiceType::OWN_PROCESS;
-
-	/// A service entry point: receives a watcher that flips to `true` when the
-	/// SCM asks the service to stop, and is responsible for exiting promptly.
-	pub type EntryFn = fn(watch::Receiver<bool>);
-
-	struct Config {
-		name: &'static str,
-		entry: EntryFn,
-	}
-
-	static CONFIG: OnceLock<Config> = OnceLock::new();
-
-	define_windows_service!(ffi_service_main, service_main);
-
-	fn service_main(_args: Vec<OsString>) {
-		let Some(config) = CONFIG.get() else {
-			return;
-		};
-
-		let (tx, rx) = watch::channel(false);
-
-		let handler = move |control_event| -> ServiceControlHandlerResult {
-			match control_event {
-				ServiceControl::Interrogate => ServiceControlHandlerResult::NoError,
-				ServiceControl::Stop | ServiceControl::Shutdown => {
-					let _ = tx.send(true);
-					ServiceControlHandlerResult::NoError
-				}
-				_ => ServiceControlHandlerResult::NotImplemented,
-			}
-		};
-
-		let Ok(status_handle) = service_control_handler::register(config.name, handler) else {
-			return;
-		};
-
-		let _ = status_handle.set_service_status(ServiceStatus {
-			service_type: SERVICE_TYPE,
-			current_state: ServiceState::Running,
-			controls_accepted: ServiceControlAccept::STOP | ServiceControlAccept::SHUTDOWN,
-			exit_code: ServiceExitCode::Win32(0),
-			checkpoint: 0,
-			wait_hint: Duration::default(),
-			process_id: None,
-		});
-
-		(config.entry)(rx);
-
-		let _ = status_handle.set_service_status(ServiceStatus {
-			service_type: SERVICE_TYPE,
-			current_state: ServiceState::Stopped,
-			controls_accepted: ServiceControlAccept::empty(),
-			exit_code: ServiceExitCode::Win32(0),
-			checkpoint: 0,
-			wait_hint: Duration::default(),
-			process_id: None,
-		});
-	}
-
-	/// Blocks until the SCM signals stop. Only call from the main thread of a
-	/// process launched by the Service Control Manager.
-	pub fn run(name: &'static str, entry: EntryFn) -> Result<()> {
-		let _ = CONFIG.set(Config { name, entry });
-		service_dispatcher::start(name, ffi_service_main)
-	}
-}
-
-#[cfg(windows)]
-pub use imp::{EntryFn, run};
-
-#[cfg(windows)]
-pub use windows_service::Error;
-
-#[cfg(not(windows))]
-pub type EntryFn = fn(());
-
-#[cfg(not(windows))]
-#[derive(Debug)]
-pub struct Error;
-
-#[cfg(not(windows))]
-impl std::fmt::Display for Error {
-	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-		write!(f, "Windows Services are only available on Windows")
-	}
-}
-
-#[cfg(not(windows))]
-impl std::error::Error for Error {}
-
-#[cfg(not(windows))]
-pub fn run(_name: &'static str, _entry: EntryFn) -> std::result::Result<(), Error> {
-	Err(Error)
-}
